requirements.txt
# ── Web Framework ──────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# ── Background Tasks ──────────────────────────────────────────
celery[redis]==5.4.0
redis==5.2.1

# ── Database ──────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
psycopg2-binary==2.9.10

# ── Configuration & Validation ────────────────────────────────
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# ── Security & Encryption ─────────────────────────────────────
cryptography==42.0.5

# ── HTTP Client ───────────────────────────────────────────────
httpx~=0.25.2

# ── Retry / Resilience ───────────────────────────────────────
tenacity==9.0.0
pybreaker==1.2.0

# ── Logging ───────────────────────────────────────────────────
loguru==0.7.3

# ── Media Processing ─────────────────────────────────────────
ffmpeg-python==0.2.0
Pillow==11.1.0

# ── Google / YouTube ─────────────────────────────────────────
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# ── AI / LLM SDKs ───────────────────────────────────────────
openai==1.59.5
anthropic==0.42.0

# ── TTS ──────────────────────────────────────────────────────
elevenlabs==1.50.5
num2words==0.5.14

# ── Telegram Bot ─────────────────────────────────────────────
python-telegram-bot==20.7

# ── Google Trends ────────────────────────────────────────────
pytrends==4.9.2

# ── Monitoring & Observability ───────────────────────────────
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==7.0.0
flower==2.0.1
